requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
authlib==1.3.0

# API & HTTP
requests==2.31.0
aiohttp==3.9.3
httpx==0.27.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
email-validator==2.1.1

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.20.0
sentry-sdk==1.40.6

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.27.0
faker==23.2.1

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.2

# CLI
click==8.1.7
rich==13.7.1
typer==0.9.0

# Rate limiting
slowapi==0.1.9

# Redis for caching (optional)
redis==5.0.1
aioredis==2.0.1

# Background tasks
celery==5.3.6
flower==2.0.1